+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Update toast for successful swap and add abort controller to fetch assets function.
 - Update the coinbase wallet id.
 - Fixed the link redirecting to staking docs.
+- Disable Core chain in output-asset on Phantom, Keplr and Leap wallet.
 
 ## [1.1.7]